requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
cffi==1.17.1
discord.py==2.4.0
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
pytz==2024.2
typing_extensions==4.12.2
yarl==1.13.1
